merge

+++ examples/processes/single/process_2.sh
@@ -9,15 +9,21 @@ yaw_min=$YAW_MIN
 yaw_max=$YAW_MAX
 range=$RANGE
 culling_r=$CULLING_R
+run_occlusion=$RUN_OCCLUSION
 
 startframe=$STARTFRAME
 endframe=$ENDFRAME
 
 for (( i=startframe+1; i<endframe; i+=6 ))
 do
-    python ./examples/conversion/convert_single.py --input ./examples/vista_traces/${file} --frame ${i} --range ${range} --process 2
-    python ./examples/basic_usage/sim_lidar.py --trace-path ./examples/vista_traces/lidar_2 --filename ${file} --frame ${i} --resolution ${resolution} --yaw-min ${yaw_min} --yaw-max ${yaw_max} --pitch-min ${pitch_min} --pitch-max ${pitch_max} --culling-r ${culling_r}
-    rm ./examples/vista_traces/lidar_2/lidar_3d*
+    if $run_occlusion
+    then
+        python ./examples/conversion/convert_single.py --input ./examples/vista_traces/${file} --frame ${i} --range ${range} --process 2 --occlusion
+        python ./examples/basic_usage/sim_lidar.py --trace-path ./examples/vista_traces/lidar_2 --filename ${file} --frame ${i} --resolution ${resolution} --yaw-min ${yaw_min} --yaw-max ${yaw_max} --pitch-min ${pitch_min} --pitch-max ${pitch_max} --culling-r ${culling_r}
+        rm ./examples/vista_traces/lidar_2/lidar_3d*
+    else
+        python ./examples/conversion/convert_single.py --input ./examples/vista_traces/${file} --frame ${i} --range ${range} --filename ${file} --yaw-min ${yaw_min} --yaw-max ${yaw_max} --pitch-min ${pitch_min} --pitch-max ${pitch_max} --process 2
+    fi
 done 
 
 exit
